+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/api/UsesEditorConfigProperties.kt
@@ -1,0 +1,52 @@
+package com.pinterest.ktlint.core.api
+
+import org.ec4j.core.model.Property
+import org.ec4j.core.model.PropertyType
+
+/**
+ * Indicates [com.pinterest.ktlint.core.Rule] uses properties loaded from `.editorconfig` file.
+ *
+ * This properties could be:
+ * - universal `.editorconfig` properties defined
+ * [here](https://github.com/editorconfig/editorconfig/wiki/EditorConfig-Properties#current-universal-properties)
+ * - universal IntelliJ IDEA properties defined
+ * [here](https://github.com/JetBrains/intellij-community/blob/master/platform/lang-api/src/com/intellij/psi/codeStyle/CommonCodeStyleSettings.java)
+ * - Kotlin specific properties defined
+ * [here](https://github.com/JetBrains/kotlin/blob/master/idea/formatter/src/org/jetbrains/kotlin/idea/core/formatter/KotlinCodeStyleSettings.java)
+ *
+ * In the best case rule should only use one property.
+ *
+ * See [com.pinterest.ktlint.core.KtLint.generateKotlinEditorconfigSection] documentation how to generate
+ * `.editorconfig` based on [com.pinterest.ktlint.core.Rule]s with this interface implementations.
+ */
+@FeatureInAlphaState
+public interface UsesEditorConfigProperties {
+
+    /**
+     * Provide a list of code style editorconfig properties, that rule uses in linting.
+     */
+    public val editorconfigProperties: List<EditorConfigProperty<*>>
+
+    /**
+     * Get the value of [EditorConfigProperty] based on loaded [EditorConfigProperties] content.
+     */
+    public fun <T> EditorConfigProperties.getEditorConfigValue(property: EditorConfigProperty<T>): T {
+        return get(property.type.name)?.getValueAs() ?: property.defaultValue
+    }
+
+    /**
+     * Supported `.editorconfig` property.
+     *
+     * @param type type of property. Could be one of default ones (see [PropertyType.STANDARD_TYPES]) or custom one.
+     * @param defaultValue default value for property if it does not exist in loaded properties.
+     */
+    public data class EditorConfigProperty<T>(
+        public val type: PropertyType<T>,
+        public val defaultValue: T
+    )
+}
+
+/**
+ * Loaded [Property]s from `.editorconfig` files.
+ */
+public typealias EditorConfigProperties = Map<String, Property>